+++ R/nc_att.R
@@ -10,7 +10,8 @@
 #' @param attribute name or index (zero based) of attribute
 #' @param ... ignored
 #'
-#' @return data frame of attribute 
+#' @return data frame of attribute with numeric id, character attribute name,
+#' character or numeric variable id or name depending on input, and attribute value. 
 #' @export
 #'
 #' @examples
@@ -24,12 +25,15 @@ nc_att <- function(x, variable, attribute, ...) {
 #' @name nc_att
 #' @export 
 #' @importFrom rlang .data
+#' @importFrom stats setNames
 nc_att.NetCDF <- function(x, variable, attribute, ...) {
+  att_info <- RNetCDF::att.inq.nc(x, variable, attribute)
+  
+## att <- structure(RNetCDF::att.get.nc(x, variable, attribute), names = att_info$name)
  att <- RNetCDF::att.get.nc(x, variable, attribute)
+  tibble::as_tibble(list(id = att_info$id, name = att_info$name, variable = variable, 
+                         value = setNames(list(att), att_info$name)))
 
- faster_as_tibble(list(attribute = attribute, variable = variable, value = list(att)))
-# structure(list(attribute = attribute, variable = variable, value = list(boom = att)), class = "data.frame")
- 
 }
 
 #' @name nc_att
@@ -67,7 +71,7 @@ nc_atts <- function(x, variable = NULL, ...) {
 #' @importFrom dplyr distinct
 #' @importFrom tibble tibble
 nc_atts.NetCDF <- function(x, variable = NULL,  ...) {
-    global <- faster_as_tibble(list(id = -1, name = "NC_GLOBAL", type = "NA_character_", 
+    global <- tibble::as_tibble(list(id = -1, name = "NC_GLOBAL", type = "NA_character_", 
                    ndims = NA_real_, dimids = NA_real_, natts = nc_inq(x)$ngatts))
   
     #vars <- nc_axes(x)
@@ -86,7 +90,7 @@ nc_atts.NetCDF <- function(x, variable = NULL,  ...) {
   var <- dplyr::bind_rows(var, global)
   #bind_rows(lapply(split(var, var$name), function(v) bind_rows(lapply(seq_len(v$natts), function(iatt) nc_att(x, v$name, iatt - 1)))))
 #bind_rows <- function(x) x
-   out <-  dplyr::bind_rows(lapply(split_fast_tibble(var, var$name), 
+   out <-  dplyr::bind_rows(lapply(split(var, var$name)[unique(var$name)], 
                      function(v) dplyr::bind_rows(lapply(seq_len(v$natts), function(iatt) nc_att(x, v$name, iatt - 1)))))
    if (!is.null(variable) && !variable %in% out$variable) stop("specified variable not found")
    vv <- variable[1]
@@ -127,13 +131,13 @@ nc_att_internal <- function(x, variable_id, attribute_id, variable_name) {
   #              attribute, as.integer(numflag), as.integer(globflag),
   #              PACKAGE = "RNetCDF")
   att <- RNetCDF::att.get.nc(x, variable_id, attribute_id)
-  faster_as_tibble(list(attribute = attribute, variable = variable_name, value = att))
+  tibble::as_tibble(list(attribute = attribute, variable = variable_name, value = att))
 }
 
 
 nc_atts_internal <- function(x, n_global_atts, variables = NULL, ...) {
 
-global <- faster_as_tibble(list(id = -1, name = "NC_GLOBAL", type = "NA_character_",
+global <- tibble::as_tibble(list(id = -1, name = "NC_GLOBAL", type = "NA_character_",
                                   ndims = NA_real_, natts = n_global_atts,
                                 dim_coord = FALSE))