+++ R/nc_grid.R
@@ -42,6 +42,11 @@ nc_grids.NetCDF <- function(x, ...) {
 # nc_axes(f, var) ## just these ones
 # nc_axis(i)  ## just one, of all ??
 
+expand_var <- function(x) {
+  nc_axes(x) %>% 
+    dplyr::inner_join(nc_dims(x), c("dimension" = "id")) %>% 
+    dplyr::inner_join(nc_vars(x), c("variable" = "name"))
+}
 
 #' @importFrom dplyr desc arrange
 #' @importFrom rlang .data
@@ -50,15 +55,20 @@ nc_grids_dimvar <- function(dimension, variable, axes) {
   if (is.null(variable) || (nrow(variable) < 1 & nrow(dimension) < 1)) return(tibble::tibble())
   native_var <- unique(axes$variable)
   shape_instances_byvar <- split(axes$dimension, axes$variable)[native_var]
-#    axes %>% 
- #   split_fast_tibble(.$variable) %>% purrr::map(function(xa) xa$dimension)
   shape_classify_byvar <- factor(unlist(lapply(shape_instances_byvar, 
                                                function(xb) paste(paste0("D", xb), collapse = ","))))
-  out <- faster_as_tibble(list(variable  = names(shape_classify_byvar), 
+  out <- tibble::as_tibble(list(variable  = names(shape_classify_byvar), 
                 grid = levels(shape_classify_byvar)[shape_classify_byvar]))
   out <-   dplyr::arrange(out, dplyr::desc(nchar(.data$grid)), .data$grid)
   ## catch the NA shapes (the scalars) and set to "-"
   out[["grid"]][is.na(out[["grid"]]) | out[["grid"]] == "DNA"] <- "S"
-  out  
+  out[["ndims"]] <- unlist(lapply(strsplit(out$grid, ","), length))
+  # out  %>% 
+  #   dplyr::group_by(.data$grid, .data$ndims) %>% 
+  #   dplyr::summarize(nvars = dplyr::n()) %>% 
+  #   dplyr::ungroup()
+  out <- tidyr::nest(out, .data$variable, .key = "variables") 
+  out$nvars <- unlist(lapply(out$variables, nrow))
+  out
 }